street-cam-classification/requirements.txt
flask
mxnet-cu101
gluoncv